+++ src/main/java/com/multi/hereevent/event/time/EventTimeDAO.java
@@ -10,7 +10,4 @@ public interface EventTimeDAO {
     EventTimeDTO getEventTimeByEventNoAndDay(int event_no,String day);
     List<EventTimeDTO> getEventTime(int event_no);
     List<String> getHolidayDays(int event_no);
-    int getReserveLimitByEventId(int event_no);
-    int getReservedCountByEventId(int event_no, String reserve_date, String reserve_time);
-    void insertReservation(Map<String, Object> reservationInfo);
 }
