+++ src/main/java/com/multi/hereevent/event/time/EventTimeDAOImpl.java
@@ -38,18 +38,4 @@ public class EventTimeDAOImpl implements EventTimeDAO {
 
     }
 
-    @Override
-    public int getReserveLimitByEventId(int event_no) {
-        return sqlSession.selectOne("com.multi.hereevent.event.time.getReserveLimitByEventId", event_no);
-    }
-
-    @Override
-    public int getReservedCountByEventId(int event_no, String reserve_date, String reserve_time) {
-        return sqlSession.selectOne("com.multi.hereevent.event.time.getReservedCountByEventId", event_no);
-    }
-
-    @Override
-    public void insertReservation(Map<String, Object> reservationInfo) {
-        sqlSession.insert("com.multi.hereevent.event.time.insertReservation", reservationInfo);
-    }
 }
