+++ src/main/java/com/multi/hereevent/event/time/EventTimeService.java
@@ -10,5 +10,4 @@ public interface EventTimeService {
     EventTimeDTO getEventTimeByEventNoAndDay(int event_no, String day);
     List<EventTimeDTO> getEventTime(int event_no);
     List<String> getHolidayDays(int event_no);
-    String makeReservation(int event_no, int member_no, String reserve_date, String reserve_time);
 }
