+++ src/main/resources/application.properties
@@ -5,38 +5,38 @@ server.servlet.context-path=/hereevent
 # MySQL Driver
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 ## db source url
-#spring.datasource.url=jdbc:mysql://db-nt04n-kr.vpc-pub-cdb.ntruss.com:3306/hereevent?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://db-o6tr6-kr.vpc-pub-cdb.ntruss.com:3306/hereevent?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
 ## db response name
-#spring.datasource.username=hereevent
+spring.datasource.username=hereevent
 ## db response password
-#spring.datasource.password=hereevent2024!
+spring.datasource.password=hereevent2024!
 # db source url
-spring.datasource.url=jdbc:mysql://localhost:3306/hereevent?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
-# db response name
-spring.datasource.username=root
-# db response password
-spring.datasource.password=1234
+#spring.datasource.url=jdbc:mysql://localhost:3306/hereevent?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
+## db response name
+#spring.datasource.username=root
+## db response password
+#spring.datasource.password=1234
 
 # MyBatis
 mybatis.mapper-locations=classpath:/mapper/*.xml
 mybatis.type-aliases-package=com.multi.hereevent.dto
 
-# response 한글 깨짐 해결
+# response \u00ED\u0095\u009C\u00EA\u00B8\u0080 \u00EA\u00B9\u00A8\u00EC\u00A7\u0090 \u00ED\u0095\u00B4\u00EA\u00B2\u00B0
 server.servlet.encoding.charset=UTF-8
 server.servlet.encoding.force=true
 
 # SMTP
-# gmail 로 보내기
+# gmail \u00EB\u00A1\u009C \u00EB\u00B3\u00B4\u00EB\u0082\u00B4\u00EA\u00B8\u00B0
 spring.mail.host=smtp.gmail.com
-# gmail 이 사용하는 포트
+# gmail \u00EC\u009D\u00B4 \u00EC\u0082\u00AC\u00EC\u009A\u00A9\u00ED\u0095\u0098\u00EB\u008A\u0094 \u00ED\u008F\u00AC\u00ED\u008A\u00B8
 spring.mail.port=587
-# 발송에 사용될 계정의 메일 주소
+# \u00EB\u00B0\u009C\u00EC\u0086\u00A1\u00EC\u0097\u0090 \u00EC\u0082\u00AC\u00EC\u009A\u00A9\u00EB\u0090\u00A0 \u00EA\u00B3\u0084\u00EC\u00A0\u0095\u00EC\u009D\u0098 \u00EB\u00A9\u0094\u00EC\u009D\u00BC \u00EC\u00A3\u00BC\u00EC\u0086\u008C
 spring.mail.username=sy0223lee@gmail.com
-# 발송에 사용될 계정의 비밀번호
+# \u00EB\u00B0\u009C\u00EC\u0086\u00A1\u00EC\u0097\u0090 \u00EC\u0082\u00AC\u00EC\u009A\u00A9\u00EB\u0090\u00A0 \u00EA\u00B3\u0084\u00EC\u00A0\u0095\u00EC\u009D\u0098 \u00EB\u00B9\u0084\u00EB\u00B0\u0080\u00EB\u00B2\u0088\u00ED\u0098\u00B8
 spring.mail.password=sriz fvpg yyaz xqvv
-# SMTP 인증을 사용
+# SMTP \u00EC\u009D\u00B8\u00EC\u00A6\u009D\u00EC\u009D\u0084 \u00EC\u0082\u00AC\u00EC\u009A\u00A9
 spring.mail.properties.mail.smtp.auth=true
-# TLS 암호화를 사용하여 이메일을 전송
+# TLS \u00EC\u0095\u0094\u00ED\u0098\u00B8\u00ED\u0099\u0094\u00EB\u00A5\u00BC \u00EC\u0082\u00AC\u00EC\u009A\u00A9\u00ED\u0095\u0098\u00EC\u0097\u00AC \u00EC\u009D\u00B4\u00EB\u00A9\u0094\u00EC\u009D\u00BC\u00EC\u009D\u0084 \u00EC\u00A0\u0084\u00EC\u0086\u00A1
 spring.mail.properties.mail.smtp.starttls.enable=true
-# TLS 암호화를 필수로 사용
+# TLS \u00EC\u0095\u0094\u00ED\u0098\u00B8\u00ED\u0099\u0094\u00EB\u00A5\u00BC \u00ED\u0095\u0084\u00EC\u0088\u0098\u00EB\u00A1\u009C \u00EC\u0082\u00AC\u00EC\u009A\u00A9
 spring.mail.properties.mail.smtp.starttls.required=true
